serial runs on workers now, other fixes, not perfect but kinda fun

+++ src/worker_pipes.ts
@@ -1,12 +1,13 @@
 
-import {WorkerService, ServiceMessage} from 'graphscript'//'../../GraphServiceRouter/index'
+import {WorkerService, ServiceMessage} from '../../GraphServiceRouter/index' //'graphscript'//
 import { WorkerInfo } from 'graphscript';
 import {WebSerial} from './serial/serialstream'
 import gsworker from './debugger.worker'
 
 export const workers = new WorkerService(); 
 
-import { WebglLinePlotUtil, WebglLinePlotProps, WebglLinePlotInfo, WebglLineProps } from '../../BrainsAtPlay_Libraries/webgl-plot-utils/webgl-plot-utils'//'webgl-plot-utils';
+import { WebglLinePlotUtil, WebglLinePlotProps, WebglLinePlotInfo, WebglLineProps } from 'webgl-plot-utils'//'../../BrainsAtPlay_Libraries/webgl-plot-utils/webgl-plot-utils'//'webgl-plot-utils';
+
 
 
 //transfer decoders
@@ -140,7 +141,7 @@ export function transferStreamAPI(worker:WorkerInfo) {
             globalThis.decoder = 'raw';
             console.log('worker: Setting up Serial', globalThis.Serial)
 
-            globalThis.Serial.getPorts().then(console.log)
+            //globalThis.Serial.getPorts().then(console.log)
             return true;
         },
         'setupSerial'
@@ -148,97 +149,76 @@ export function transferStreamAPI(worker:WorkerInfo) {
 
     transferFunction(
         worker,
-        function openPort(self, origin, settings:SerialOptions & { usbVendorId:number, usbProductId:number, pipeTo?:string|{route:string, _id:string}, frequency?:number }) {
-            globalThis.Serial.getPorts().then((ports)=>{
-                const WorkerService = self.graph as WorkerService;
-                if(!globalThis.Serial) WorkerService.run('setupSerial');
+        function openPort(self, origin, settings:SerialOptions & { usbVendorId:number, usbProductId:number, pipeTo?:string|{route:string, _id:string, extraArgs:any[]}, frequency?:number, buffering?:{searchBytes:Uint8Array} }) {
+            
+            const WorkerService = self.graph as WorkerService;
+            if(!globalThis.Serial) WorkerService.run('setupSerial');
+            return new Promise((res,rej) => {
+                globalThis.Serial.getPorts().then((ports)=>{
+        
+                    const Serial = globalThis.Serial as WebSerial;
     
-                const Serial = globalThis.Serial as WebSerial;
-
-                let port = ports.find((port)=>{
-                    return port.getInfo().usbVendorId === settings.usbVendorId && port.getInfo().usbProductId === settings.usbProductId;
-                });
-                if(port) {
-                    Serial.openPort(port, settings).then(() => {
-                        const stream = Serial.createStream({
-                            port, 
-                            frequency:settings.frequency ? settings.frequency : 10,
-                            ondata: (value:ArrayBuffer) => { 
-                                if(globalThis.decoder) value = WorkerService.run(globalThis.decoder, value); //run the decoder if set on this thread, else return the array buffer result raw or pipe to another thread
-    
-                                if(stream.settings.pipeTo) {
-                                    if(typeof stream.settings.pipeTo === 'string')
-                                        WorkerService.transmit(value, stream.settings.pipeTo, (value instanceof ArrayBuffer || (value as any).constructor?.name.indexOf('Array') > 0) ? [value] as any : undefined);
-                                    //we can subscribe on the other end to this worker output by id
-                                    else if (stream.settings.pipeTo?.route) {
-                                        WorkerService.transmit({route:stream.settings.pipeTo.route, args:value }, stream.settings.pipeTo._id, (value instanceof ArrayBuffer || (value as any).constructor?.name.indexOf('Array') > 0) ? [value] as any : undefined);
-                                    }
-                                } else {
-                                    WorkerService.transmit(value, origin, (value instanceof ArrayBuffer || (value as any).constructor?.name.indexOf('Array') > 0) ? [value] as any : undefined);
-                                    //we can subscribe on the other end to this worker output by id
-                                }
-                            }
-                        });
-                        stream.settings = settings; //save the settings 
-    
-                        Serial.readStream(stream);
+                    let port = ports.find((port)=>{
+                        return port.getInfo().usbVendorId === settings.usbVendorId && port.getInfo().usbProductId === settings.usbProductId;
                     });
-                } else {
-                    return false;
-                }
+                    if(port) {
+                        Serial.openPort(port, settings).then(() => {
+                            const stream = Serial.createStream({
+                                port, 
+                                frequency:settings.frequency ? settings.frequency : 10,
+                                buffering:settings.buffering,
+                                ondata: (value:Uint8Array) => { 
+                                    //if(globalThis.decoder) value = WorkerService.run(globalThis.decoder, value); //run the decoder if set on this thread, else return the array buffer result raw or pipe to another thread
+                                    //console.log(value);
+                                    if(stream.settings.pipeTo) {
+                                        if(typeof stream.settings.pipeTo === 'string')
+                                            WorkerService.transmit(value, stream.settings.pipeTo, [value.buffer] as any);
+                                        //we can subscribe on the other end to this worker output by id
+                                        else if (stream.settings.pipeTo?.route) {
+                                            let args:any = value;
+                                            if(stream.settings.pipeTo.extraArgs) args = [value, ...stream.settings.pipeTo.extraArgs];
+                                            WorkerService.transmit({route:stream.settings.pipeTo.route, args }, stream.settings.pipeTo._id,  [value.buffer] as any);
+                                        }
+                                    } else {
+                                        WorkerService.transmit(value, origin, [value.buffer] as any);
+                                        //we can subscribe on the other end to this worker output by id
+                                    }
+                                }
+                            });
+                            stream.settings = settings; //save the settings 
+        
+                            Serial.readStream(stream);
+        
+                            console.log(stream);
+
+                            res({
+                                _id:stream._id,
+                                settings,
+                                info:stream.info
+                            })
+                        });
+                    } else {
+                        rej(false);
+                    }
+                })
             })
+            
         },
         'openPort'
     )
-    transferFunction(
-        worker,
-        function startSerialStream(self, origin, settings:SerialOptions & { usbVendorId:number, usbProductId:number, pipeTo?:string|{route:string, _id:string}, frequency?:number }) {
-
-            const WorkerService = self.graph as WorkerService;
-            if(!globalThis.Serial) WorkerService.run('setupSerial');
-
-            const Serial = globalThis.Serial as WebSerial;
-
-            Serial.requestPort(settings.usbVendorId, settings.usbProductId).then((port) => {
-                Serial.openPort(port, settings).then(() => {
-                    const stream = Serial.createStream({
-                        port, 
-                        frequency:settings.frequency ? settings.frequency : 10,
-                        ondata: (value:ArrayBuffer) => { 
-                            if(globalThis.decoder) value = WorkerService.run(globalThis.decoder, value); //run the decoder if set on this thread, else return the array buffer result raw or pipe to another thread
-
-                            if(stream.settings.pipeTo) {
-                                if(typeof stream.settings.pipeTo === 'string')
-                                    WorkerService.transmit(value, stream.settings.pipeTo, (value instanceof ArrayBuffer || (value as any).constructor?.name.indexOf('Array') > 0) ? [value] as any : undefined);
-                                //we can subscribe on the other end to this worker output by id
-                                else if (stream.settings.pipeTo?.route) {
-                                    WorkerService.transmit({route:stream.settings.pipeTo.route, args:value }, stream.settings.pipeTo._id, (value instanceof ArrayBuffer || (value as any).constructor?.name.indexOf('Array') > 0) ? [value] as any : undefined);
-                                }
-                            } else {
-                                WorkerService.transmit(value, origin, (value instanceof ArrayBuffer || (value as any).constructor?.name.indexOf('Array') > 0) ? [value] as any : undefined);
-                                //we can subscribe on the other end to this worker output by id
-                            }
-                        }
-                    });
-                    stream.settings = settings; //save the settings 
-
-                    Serial.readStream(stream);
-                    
-                });
-            }).catch(console.error);
-        
-            return true;
-        },
-        'startSerialStream'
-    );
+   
     transferFunction(
         worker,
         function closeStream(self, origin, streamId) {
-            const Serial = globalThis.Serial as WebSerial;
+            return new Promise((res,rej) => {
 
-            Serial.closeStream(Serial.streams[streamId]);
+                const Serial = globalThis.Serial as WebSerial;
 
-            return true;
+                Serial.closeStream(Serial.streams[streamId]).then((resolved) => {
+                    res(resolved);
+                }).catch(rej);
+            });
+
         },
         'closeStream'
     );
@@ -246,12 +226,26 @@ export function transferStreamAPI(worker:WorkerInfo) {
         worker,
         function writeStream(self, origin, streamId, message:any) {
 
-            globalThis.Serial.writeStream(globalThis.Serial.streams[streamId], message).then(console.log);
+            (globalThis.Serial as WebSerial).writeStream(globalThis.Serial.streams[streamId], message).then(console.log);
 
             return true;
         },
         'writeStream'
     );
+    transferFunction(
+        worker,
+        function updateStreamSettings(self,origin, streamId:string, settings:any) {
+            if(globalThis.Serial?.streams[streamId]) {
+                for(const key in settings) {
+                    if(typeof settings[key] === 'object') {
+                        Object.assign(globalThis.Serial.streams[streamId].settings[key], settings[key]);
+                    }
+                    else globalThis.Serial.streams[streamId][key] = settings[key];
+                }
+            }
+        },
+        'updateStreamSettings'
+    )
 }
 //make a serial stream worker and a decoder worker separately,
 //workers.establishMessageChannel(worker1, worker2);
@@ -279,21 +273,20 @@ export function initWorkerChart(
     const chart = document.createElement('canvas');
 
     const devicePixelRatio = window.devicePixelRatio || 1;
-    (chart as any).width = parentDiv.clientWidth;
-    (chart as any).height = parentDiv.clientHeight;
+    (chart as any).width = parentDiv.clientWidth ? parentDiv.clientWidth : window.innerWidth;
+    (chart as any).height = parentDiv.clientHeight ? parentDiv.clientHeight : 450;
     chart.style.width = '100%';
     chart.style.height = '100%';
 
     const overlay = document.createElement('canvas');
 
-    overlay.style.position = 'absolute';
-    overlay.style.zIndex = '10';
-    (overlay as any).width = parentDiv.clientWidth;
-    (overlay as any).height = parentDiv.clientHeight;
+    (overlay as any).width = parentDiv.clientWidth ? parentDiv.clientWidth : window.innerWidth;
+    (overlay as any).height = parentDiv.clientHeight ? parentDiv.clientHeight : 450;
     overlay.style.left = parentDiv.offsetLeft + 'px';
     overlay.style.top = parentDiv.offsetTop + 'px';
-    overlay.style.width = parentDiv.clientWidth + 'px';
-    overlay.style.height = parentDiv.clientHeight + 'px';
+    overlay.style.width = (parentDiv.clientWidth ? parentDiv.clientWidth : window.innerWidth) + 'px';
+    overlay.style.height = (parentDiv.clientHeight ? parentDiv.clientHeight : 450) + 'px';
+    overlay.style.transform = `translateY(-${overlay.height}px)`;
 
     plotDiv.appendChild(chart);
     plotDiv.appendChild(overlay);
@@ -335,14 +328,23 @@ export function initWorkerChart(
 }
 
 
-export function createStreamRenderPipeline() {
+export function createStreamRenderPipeline(dedicatedSerialWorker=false) {
     let streamworker = workers.addWorker({url:gsworker}) as WorkerInfo;
     let chartworker = workers.addWorker({url:gsworker}) as WorkerInfo;
+    let serialworker, decoderPort, chartPort;
+
+    if(dedicatedSerialWorker) {
+        serialworker = workers.addWorker({url:gsworker}) as WorkerInfo;
+        transferStreamAPI(serialworker);
+        decoderPort = workers.establishMessageChannel(serialworker.worker, streamworker.worker); //returns the id of the port so we can orchestrate port communication
+    
+        serialworker.post('setupSerial');
+    }   
 
     transferStreamAPI(streamworker);
     transferChartCommands(chartworker);
 
-    let portId = workers.establishMessageChannel(streamworker.worker, chartworker.worker); //returns the id of the port so we can orchestrate port communication
+    chartPort = workers.establishMessageChannel(streamworker.worker, chartworker.worker); //returns the id of the port so we can orchestrate port communication
 
     // initWorkerChart(
     //     chartworker
@@ -352,7 +354,6 @@ export function createStreamRenderPipeline() {
         streamworker,
         function decodeAndPassToChart(self, origin, data:any, chartPortId:string) {
             let decoded = self.graph.run('decode',data);
-            //console.log('decoded', decoded)
             if(decoded) self.graph.workers[chartPortId].send(
                 {
                     route:'updateChartData',
@@ -360,6 +361,7 @@ export function createStreamRenderPipeline() {
                 }//,
                 //chartPortId
             );
+            //console.log(decoded, self.graph)
             return decoded;
         },
         'decodeAndPassToChart'
@@ -369,11 +371,17 @@ export function createStreamRenderPipeline() {
     //for serial we need to tell the serial port/decoder worker to do that instead of the main thread,
     // and proxy the serial port controls
 
-    return {
+    let result:any = {
         streamworker,
         chartworker,
-        portId
+        chartPort, //stream --> chart worker message port
     };
+    if(serialworker) {
+        result.serialworker = serialworker;
+        result.decoderPort = decoderPort; //serial --> stream worker message port
+    }
+
+    return result;
 }
 
 //after calling createStreamRenderPipeline
@@ -401,9 +409,10 @@ export function initWorkerSerialStream(
 }
 
 
-export function cleanupWorkerStreamPipeline(streamworker, chartworker, plotDiv?:HTMLElement) {
-    workers.terminate(streamworker.worker);
-    workers.terminate(chartworker.worker);
+export function cleanupWorkerStreamPipeline(streamworker, chartworker, plotDiv?:HTMLElement, serialworker?) {
+    if(streamworker) workers.terminate(streamworker.worker);
+    if(chartworker) workers.terminate(chartworker.worker);
+    if(serialworker) workers.terminate(serialworker.worker);
 
     if(plotDiv) plotDiv.remove();   
 }
